+++ tests/testthat/test_inference.R
@@ -408,12 +408,6 @@ test_that("mcmc errors for invalid parallel plans", {
   # reset warning setting
   withr::defer(future::plan(op))
 
-  # handle handle forks, so only accept multisession, or multi session clusters
-  suppressWarnings(future::plan(future::multiprocess))
-  expect_snapshot_error(
-    mcmc(m, verbose = FALSE)
-  )
-
   future::plan(future::multicore)
   expect_snapshot_error(
     mcmc(m, verbose = FALSE)