+++ NEWS.md
@@ -1,3 +1,5 @@
+# greta (development version)
+
 # greta 0.4.3
 
 ## Features